+++ lib/jwtUtils.js
@@ -1,7 +1,7 @@
 import dotenv from 'dotenv';
 import jwt from 'jsonwebtoken';
 import consoleLogger from './consoleLogger.js';
-import { insertToken, selectToken } from '../models/accessTokenTable.js';
+import { insertToken, selectToken } from '../models/accessTokenRedis.js';
 
 dotenv.config();
